pyproject.toml
[tool.poetry]
name = "streamlistapp"
version = "0.1.0"
description = ""
authors = ["Elephantatech <vivekm.mistry@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
streamlit-carousel = "^1.1.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"